cmd/compile: Fix nilcheck generation for Aix
Nilcheck must be changed on Aix because a load of 0x0 address won't raise a SIGSEGV.
Therefore, a nilcheck must compare the current pointer address with 0x0 and branch to a function
which will panic: runtime.NilDereferencePanic

Wrappers generation must also be changed because of the presence of a BL in Aix nilcheck.
The previous generation used a RETJMP but it implies that the current function have no frame.
With the BL generated, a frame is mandatory, so RETJMP must be disabled.