WIP: Support GOOS=os400
Anywhere we check GOOS for aix, also check for os400.

Also adjust build lines so that any that are built on AIX also get
built on IBM i and any that are not built on AIX are not built on IBM i.
For now, we ignore any AIX-specific files as that has been split in to a
separate commit.

TODO: These commits should be merged. Also, some tests may need to be
disabled for IBM i and the build lines adjusted accordingly.